In-place sed for both GNU sed and BSD/macOS sed
For GNU sed, the suffix for the -i option is optional.
But OS X ships with BSD sed, where the suffix is mandatory.

The trick proposed here make the command works with both implementations.
It would not work with -i'' only, because to sed it is indistinguishable from
just -i, because the shell effectively removes the empty quotes and passes just
-i.

With (effectively) just -i specified, it is the next argument that is
interpreted as the option-argument

Note :
it is not portable in a POSIX sense, because the -i option is not
POSIX-compliant.